Fix the detection of code blocks in the editor
The original logic of the code blocks does not understand the concept of an empty highlighter value to indicate a server-side detection.

We can work around this limitation by providing the command (and the button for the active state) with the proper value. A `value` of `false` means that there is no code block and all other cases are then teated as “there is a code block“.

See https://www.woltlab.com/community/thread/301297-aktivstatus-code-block-button-fehlt/